シェアされたユーザのみ画像が見えるよう修正
