+++ lambda/app/model/memo.py
@@ -11,6 +11,7 @@ from common_headers import *
 from my_common import *
 from dynamo_utility import *
 from model.share import *
+from service.user import *
 
 db_resource = boto3.resource("dynamodb")
 db_client = boto3.client("dynamodb", region_name='ap-northeast-1')
@@ -405,11 +406,6 @@ def delete_share_setting_multi_by_memo_id(memo_ids) -> bool:
         return False
     return False
 
-def check_is_in_share_target(user_id: str, share_targets: str) -> bool:
-    if not share_targets or not user_id:
-        return False
-    target_users = share_targets.replace(' ', '').split(',')
-    return user_id in target_users
 
 '''
 @param list user_data get_user_data_by_uuid等から取得したもの. 主にログイン中のユーザ
